US1: Implementera visning av minst 3 studenters namn
