+++ index.html
@@ -10,6 +10,6 @@
     <h1>Studentkatalog</h1>
     <div id="student-list"></div>
     
-    <script type="module" src="src/main.js"></script>
+    <script type="module" src="dist/main.js" defer></script>
 </body>
 </html>
